% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R
\name{add_mh_flank}
\alias{add_mh_flank}
\title{get mh value in flank region of SurVirus result (result of get_seq())}
\usage{
add_mh_flank(result_rel, len, unite = TRUE)
}
\arguments{
\item{result_rel}{SurVirus result (result of get_seq())}

\item{len}{flanking region size}

\item{unite}{If TRUE, calculate the lengths of each microsome separately, 
otherwise calculate the sum of microsome lengths}
}
\value{
dataframe
}
\description{
get mh value in flank region of SurVirus result (result of get_seq())
}
\examples{
data(BSgenomehpv16)
SurVirus_dir <- system.file(file.path("extdata", "survirus_result"), 
    package = "mhAnalysis")
bam_dir <- file.path(SurVirus_dir, "readsx")

result_t1 <- read.table(file.path(SurVirus_dir, "results.t1.txt"), sep = " ", 
    header = FALSE, fill = TRUE)
results <- read.table(file.path(SurVirus_dir, "results.txt"), sep = " ", 
    header = FALSE, fill = TRUE)
rownames(results) <- paste("ID", results[, 1], sep = "=")
results <- results[result_t1[, 1], ]
result_rel <- add_strand(results, bam_dir, result_t1)
result_rel <- get_real_loc(result_rel, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
result_rel <- get_seq(result_rel, len = 10, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
result_rel <- add_mh_flank(result_rel, len = 5)
}
