% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{add_strand}
\alias{add_strand}
\title{Add strand information}
\usage{
add_strand(results, bam_dir, result_t1)
}
\arguments{
\item{results}{dataframe of results.txt in SurVirus software}

\item{bam_dir}{The directory of readsx folder (contains bam files) of}

\item{result_t1}{result_t1.txt of SurVirus software}
}
\value{
datafrmae contaions strand information
}
\description{
Add strand information
}
\examples{
data(BSgenomehpv16)
SurVirus_dir <- system.file(file.path("extdata", "survirus_result"), 
    package = "mhAnalysis")
bam_dir <- file.path(SurVirus_dir, "readsx")

result_t1 <- read.table(file.path(SurVirus_dir, "results.t1.txt"), sep = " ", 
    header = FALSE, fill = TRUE)
results <- read.table(file.path(SurVirus_dir, "results.txt"), sep = " ", 
    header = FALSE, fill = TRUE)
rownames(results) <- paste("ID", results[, 1], sep = "=")
results <- results[result_t1[, 1], ]
result_rel <- add_strand(results, bam_dir, result_t1)
result_rel <- get_real_loc(result_rel, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
result_rel <- get_seq(result_rel, len = 10, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
}
