% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSgenomehpv16}
\alias{BSgenomehpv16}
\title{A BSgenome object of HPV16}
\format{
A BSgenome
}
\usage{
BSgenomehpv16
}
\value{
A BSgenome object of HPV16
}
\description{
Full genomic sequences for HPV16
}
\keyword{datasets}
