% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_comple}
\alias{get_comple}
\title{Get complementary sequences}
\usage{
get_comple(x)
}
\arguments{
\item{x}{a sequence contains A T G C N}
}
\value{
a complementary sequence contains A T G C N
}
\description{
Get complementary sequences
}
\examples{
get_comple("ATTCCGNNGCC")
}
