% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R
\name{get_mh}
\alias{get_mh}
\title{Get microhomology(mh) length}
\usage{
get_mh(seq_df, unite = TRUE)
}
\arguments{
\item{seq_df}{a dataframe of three rows, each containing the host sequence,
integration sequence, and virus sequence}

\item{unite}{If TRUE, calculate the lengths of each microsome separately, otherwise calculate the sum of microsome lengths}
}
\value{
a numeric of mh length
}
\description{
Get microhomology(mh) length
}
\examples{
seq_df <- data.frame(host = c("A", "T", "G", "G", "C", "T", "A", "A"),
                     inte = c("A", "T", "G", "G", "A", "T", "A", "C"),
                     virus= c("C", "A", "T", "T", "A", "T", "A", "C"))
seq_df <- t(seq_df) |> as.data.frame()
get_mh(seq_df)
}
