% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_r}
\alias{get_r}
\title{get_r}
\usage{
get_r(y)
}
\arguments{
\item{y}{flag of bam/sam file, a vector.}
}
\value{
a vector of character, shows the reads are "read1" or "read2"
}
\description{
get_r
}
\examples{
get_r(c(68,74))
}
