% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_rev}
\alias{get_rev}
\title{Get reverse sequences}
\usage{
get_rev(x)
}
\arguments{
\item{x}{a sequence contains A T G C N}
}
\value{
a reverse sequence
}
\description{
Get reverse sequences
}
\examples{
get_rev("ATTCCGNNGCC")
}
