% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_seq}
\alias{get_seq}
\title{Get sequence of SurVirus result, and update the location(start and end) of
host sequence and virus sequence}
\usage{
get_seq(result_rel, len = 10, BSgenome_host, BSgenome_virus)
}
\arguments{
\item{result_rel}{result of get_real_loc()}

\item{len}{flank region size}

\item{BSgenome_host}{BSgenome object of host}

\item{BSgenome_virus}{BSgenome object of virus}
}
\value{
dataframe
}
\description{
Get sequence of SurVirus result, and update the location(start and end) of
host sequence and virus sequence
}
\examples{
data(BSgenomehpv16)
SurVirus_dir <- system.file(file.path("extdata", "survirus_result"), package = "mhAnalysis")
bam_dir <- file.path(SurVirus_dir, "readsx")

result_t1 <- read.table(file.path(SurVirus_dir, "results.t1.txt"), sep = " ", header = FALSE, fill = TRUE)
results <- read.table(file.path(SurVirus_dir, "results.txt"), sep = " ", header = FALSE, fill = TRUE)
rownames(results) <- paste("ID", results[, 1], sep = "=")
results <- results[result_t1[, 1], ]
result_rel <- add_strand(results, bam_dir, result_t1)
result_rel <- get_real_loc(result_rel, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
result_rel <- get_seq(result_rel, len = 10, BSgenome_host=BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, 
    BSgenome_virus=BSgenomehpv16)
}
