% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_mh_flank}
\alias{add_mh_flank}
\title{get mh value in flank region of SurVirus result (result of get_seq())}
\usage{
add_mh_flank(result_rel, len, unite = TRUE)
}
\arguments{
\item{result_rel}{SurVirus result (result of get_seq())}

\item{len}{flanking region size}

\item{unite}{If TRUE, calculate the lengths of each microsome separately, otherwise calculate the sum of microsome lengths}
}
\value{
dataframe
}
\description{
get mh value in flank region of SurVirus result (result of get_seq())
}
\examples{
\dontrun{
bidui_dir <- "SurVirus_result"
bam_dir <- "SurVirus_result/readsx"

result_t1 <- fread(file.path(bidui_dir, "results.t1.txt"), sep = " ", header = FALSE, fill = TRUE)
results <- fread(file.path(bidui_dir, "results.txt"), sep = " ", header = FALSE, fill = TRUE)
class(results) <- class(result_t1) <- "data.frame"
rownames(results) <- paste("ID", results[, 1], sep = "=")
results <- results[result_t1[, 1], ]
result_rel <- add_strand(results)
result_rel <- get_real_loc(result_rel)
result_rel <- get_seq(result_rel)
result_rel <- add_mh_flank(result_rel, len = 5)
}
}
