% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_r}
\alias{get_r}
\title{Use the flag value to detect which of the two ends the sequence belongs to}
\usage{
get_r(y)
}
\arguments{
\item{y}{flag of bam/sam file, a vector.}
}
\value{
a vector of character, shows the reads are "read1" or "read2"
}
\description{
Use the flag value to detect which of the two ends the sequence belongs to
}
\examples{
get_r(c(68,74))
}
