% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_real_loc}
\alias{get_real_loc}
\title{Get the true insertion site of the virus}
\usage{
get_real_loc(result_rel, BSgenome_host, BSgenome_virus)
}
\arguments{
\item{result_rel}{dataframe of results.txt in SurVirus software,
which contains strand information (result of add_strand())}

\item{BSgenome_host}{BSgenome object of host}

\item{BSgenome_virus}{BSgenome object of virus}
}
\value{
data.frame
}
\description{
Get the true insertion site of the virus
}
\examples{
\dontrun{
bidui_dir <- "SurVirus_result"
bam_dir <- "SurVirus_result/readsx"

result_t1 <- fread(file.path(bidui_dir, "results.t1.txt"), sep = " ", header = FALSE, fill = TRUE)
results <- fread(file.path(bidui_dir, "results.txt"), sep = " ", header = FALSE, fill = TRUE)
class(results) <- class(result_t1) <- "data.frame"
rownames(results) <- paste("ID", results[, 1], sep = "=")
results <- results[result_t1[, 1], ]
result_rel <- add_strand(results)
result_rel <- get_real_loc(result_rel)
}
}
